Extracted `trait Source` and `impl`s into `crate::source`.
